Move zone conditions to the zone integration
